equipo1Login/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-decouple==3.8
mysqlclient==2.2.8
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
